feature(zod): moved `q.slug` into zod, since that's the only place it's useful
